updated to 1280x720p, imported cpp files, changed vector2d to int

+++ Src/Vector2D.cpp
@@ -6,7 +6,7 @@ Vector2D::Vector2D()
 	y = 0.0f;
 }
 
-Vector2D::Vector2D(float x, float y)
+Vector2D::Vector2D(int x, int y)
 {
 	this->x = x;
 	this->y = y;
